backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.5
requests>=2.28.1
tenacity>=8.1.0
sentry-sdk>=2.13.0
jinja2>=3.1.2
pydantic>=2.0,<2.7
pydantic-settings>=2.0.3
httpx>=0.23.1
setuptools>=65.6.3
google-generativeai>=0.8.5